+++ app/models/carto/api_key.rb
@@ -39,6 +39,14 @@ module Carto
     TYPE_MASTER = 'master'.freeze
     TYPE_DEFAULT_PUBLIC = 'default'.freeze
 
+    MASTER_NAME         = 'master-api-key'.freeze
+    DEFAULT_PUBLIC_NAME = 'default-public-api-key'.freeze
+
+    API_SQL       = 'sql'.freeze
+    API_MAPS      = 'maps'.freeze
+    API_IMPORT    = 'import'.freeze
+    API_ANALYSIS  = 'analysis'.freeze
+
     VALID_TYPES = [TYPE_REGULAR, TYPE_MASTER, TYPE_DEFAULT_PUBLIC].freeze
 
     self.inheritance_column = :_type
@@ -47,6 +55,7 @@ module Carto
 
     before_create :create_token
     before_create :create_db_config
+    before_validation :check_master_key
 
     serialize :grants, Carto::CartoJsonSymbolizerSerializer
     validates :grants, carto_json_symbolizer: true, api_key_grants: true, json_schema: true
@@ -61,6 +70,7 @@ module Carto
     after_destroy :remove_from_redis
 
     validates :type, inclusion: { in: VALID_TYPES }
+    validate :valid_name_for_type
 
     attr_writer :redis_client
 
@@ -100,6 +110,16 @@ module Carto
       end while self.class.exists?(token: token)
     end
 
+    def master?
+      type == TYPE_MASTER
+    end
+
+    def valid_name_for_type
+      if !master? && name == MASTER_NAME
+        errors.add(:name, "api_key name cannot be #{MASTER_NAME}")
+      end
+    end
+
     private
 
     PASSWORD_LENGTH = 40
@@ -127,14 +147,25 @@ module Carto
       table_permissions
     end
 
+    def current_user
+      user || Carto::User[user_id]
+    end
+
     def create_db_config
-      begin
-        self.db_role = Carto::DB::Sanitize.sanitize_identifier("#{user.username}_role_#{SecureRandom.hex}")
-      end while self.class.exists?(db_role: db_role)
-      self.db_password = SecureRandom.hex(PASSWORD_LENGTH / 2) unless db_password
+      if master?
+        self.db_role = current_user.database_username
+        self.db_password = current_user.database_password
+      else
+        begin
+          self.db_role = Carto::DB::Sanitize.sanitize_identifier("#{user.username}_role_#{SecureRandom.hex}")
+        end while self.class.exists?(db_role: db_role)
+        self.db_password = SecureRandom.hex(PASSWORD_LENGTH / 2) unless db_password
+      end
     end
 
     def setup_db_role
+      return if master?
+
       db_run("CREATE ROLE \"#{db_role}\" NOSUPERUSER NOCREATEDB LOGIN ENCRYPTED PASSWORD '#{db_password}'")
       db_run("GRANT \"#{user.service.database_public_username}\" TO \"#{db_role}\"")
       db_run("ALTER ROLE \"#{db_role}\" SET search_path TO #{user.db_service.build_search_path}")
@@ -153,6 +184,8 @@ module Carto
     end
 
     def drop_db_role
+      return if master?
+
       revoke_privileges
       db_run("DROP ROLE \"#{db_role}\"")
     end
@@ -205,6 +238,17 @@ module Carto
     def grant_aux_write_privileges_for_schema(s)
       db_run("GRANT USAGE ON SCHEMA \"#{s}\" TO \"#{db_role}\"")
       db_run("GRANT USAGE, SELECT ON ALL SEQUENCES IN SCHEMA \"#{s}\" TO \"#{db_role}\"")
+    end
+
+    def check_master_key
+      return unless master?
+      raise Carto::UnprocesableEntityError.new("Duplicate master API Key") if exists_master_key?(user_id)
+      self.name = MASTER_NAME
+      self.grants = [{ type: "apis", apis: [API_SQL, API_MAPS, API_IMPORT, API_ANALYSIS] }]
+    end
+
+    def exists_master_key?(user_id)
+      Carto::ApiKey.exists?(id: user_id, type: TYPE_MASTER)
     end
   end
 end