+++ spec/requests/warden_spec.rb
@@ -6,12 +6,12 @@ describe 'Warden :auth_api Strategy' do
   include HelperMethods
 
   def generate_api_key_url
-    api_keys_url(user_domain: @user1.username)
+    api_keys_url(user_domain: @user_no_api_keys.username)
   end
 
   before :all do
     @auth_api_feature_flag = FactoryGirl.create(:feature_flag, name: 'auth_api', restricted: false)
-    @master_api_key = FactoryGirl.create(:api_key_apis, user_id: @user1.id, type: 'master')
+    @master_api_key = FactoryGirl.create(:master_api_key, user_id: @user_no_api_keys.id)
   end
 
   after :all do
@@ -21,7 +21,7 @@ describe 'Warden :auth_api Strategy' do
 
   it 'authenticates with header' do
     get_json generate_api_key_url, {}, http_json_headers.merge(
-      'Authorization' => 'Basic ' + Base64.encode64("#{@user1.username}:#{@master_api_key.token}")
+      'Authorization' => 'Basic ' + Base64.encode64("#{@user_no_api_keys.username}:#{@master_api_key.token}")
     ) do |response|
       response.status.should eq 200
     end
@@ -35,7 +35,7 @@ describe 'Warden :auth_api Strategy' do
 
   it 'returns 401 if header is malformed' do
     get_json generate_api_key_url, {}, http_json_headers.merge(
-      'Authorization' => 'Basicss ' + Base64.encode64("#{@user1.username}:#{@master_api_key.token}")
+      'Authorization' => 'Basicss ' + Base64.encode64("#{@user_no_api_keys.username}:#{@master_api_key.token}")
     ) do |response|
       response.status.should eq 401
     end
@@ -43,7 +43,7 @@ describe 'Warden :auth_api Strategy' do
 
   it 'returns 401 if base64 is malforemd' do
     get_json generate_api_key_url, {}, http_json_headers.merge(
-      'Authorization' => 'Basic ' + "asda2" + Base64.encode64("#{@user1.username}:#{@master_api_key.token}")
+      'Authorization' => 'Basic ' + "asda2" + Base64.encode64("#{@user_no_api_keys.username}:#{@master_api_key.token}")
     ) do |response|
       response.status.should eq 401
     end