+++ NEWS.md
@@ -113,6 +113,7 @@ ion for time-series (#12670)
 * User accounts in locked state returns 404 for resources like maps or visualizations and redirection for private endpoints (#13030)
 * Auth API
   * Keys creation (#13170)
+  * Create master API key on user creation (#13172)
   * Keys destruction (#13171)
   * Token regeneration (#13321)
   * Keys listing (#13327)